Space and planets transitions
A meteor is now fired after warning on the planet.
Game now transitions between space and the planet.
Number of meteors if ship is destroyed in space does nto start over anymore. It uses a lower number. This formula may need to be updated.
Stopped pausing & resuming the game timeline when generating meteors. There seems to be a bug in doing so that prevented the single meteor on the planet from tweening.
Some other minor adjustments such as easing.